+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/OnlineAsyncTaskAccelByteLogin.cpp
@@ -435,6 +435,15 @@ void FOnlineAsyncTaskAccelByteLogin::OnLoginSuccess()
 	UserInterface->AddOnQueryUserProfileCompleteDelegate_Handle(LoginUserNum, FOnQueryUserProfileCompleteDelegate::CreateThreadSafeSP(UserInterface.Get(), &FOnlineUserAccelByte::PostLoginBulkGetUserProfileCompleted));
 	UserInterface->QueryUserInfo(LoginUserNum, { UserId.ToSharedRef() });
 
+	// If we are using V2 sessions, send a request to update stored platform data in the session service for native sync and crossplay
+#if AB_USE_V2_SESSIONS
+	FOnlineSessionV2AccelBytePtr SessionInterface = nullptr;
+	if (FOnlineSessionV2AccelByte::GetFromSubsystem(Subsystem, SessionInterface))
+	{
+		SessionInterface->InitializePlayerAttributes(UserId.ToSharedRef().Get());	
+	}
+#endif
+
 	CompleteTask(EAccelByteAsyncTaskCompleteState::Success);
 	AB_OSS_ASYNC_TASK_TRACE_END(TEXT(""));
 }